requirements.txt
actionlib==1.11.13
angles==1.9.11
astro-pi==1.1.5
attrs==15.2.0
autobahn==0.10.3
backports.ssl-match-hostname==3.4.0.2
base-local-planner==1.14.4
bondpy==1.8.1
bzr==2.7.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
Cap1xxx==0.1.2
catkin==0.7.11
catkin-pkg==0.4.6
catkin-pkg-modules==0.4.6
chardet==2.3.0
codebug-tether==0.9.1
configobj==5.0.6
controller-manager==0.13.3
controller-manager-msgs==0.13.3
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.9.0
defer==1.0.6
defusedxml==0.4.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.12
drumhat==0.0.5
dynamic-reconfigure==1.5.49
ecdsa==0.13
empy==3.3.2
enum34==1.1.2
envirophat==0.0.6
ExplorerHAT==0.4.2
future==0.16.0
futures==3.0.5
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.10
gennodejs==2.0.1
genpy==0.6.7
gpiozero==1.3.1
httplib2==0.9.1
idna==2.0
image-geometry==1.12.8
interactive-markers==1.11.3
ipaddress==1.0.16
iso8601==0.1.12
keyring==7.3
laser-geometry==1.6.4
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lz4==0.7.0
matplotlib==1.5.1
mercurial==3.7.3
message-filters==1.12.13
microdotphat==0.1.3
mote==0.0.3
motephat==0.0.2
mpi4py==1.3.1
mpu6050-raspberrypi==1.0.3.4
msgpack-python==0.4.6
netifaces==0.10.4
nose==1.3.7
numpy==1.11.0
oauth==1.0.1
PAM==0.4.2
pantilthat==0.0.2
paramiko==1.16.0
pianohat==0.0.5
picamera==1.12
piglow==1.2.4
pigpio==1.30
Pillow==3.1.2
pluginlib==1.11.3
psutil==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.0.29
pygame===1.9.1release
Pygments==2.1
pygobject==3.20.0
pygpgme==0.3
PyMySQL==0.7.2
pynmea2==1.12.0
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyserial==3.4
pysqlite==2.7.0
Pyste==0.9.10
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.4.2
python-debian==0.1.27
python-engineio==2.2.0
python-qt-binding==0.3.3
python-snappy==0.5
python-socketio==2.0.0
pytz==2014.10
pyxdg==0.25
PyYAML==3.12
qt-dotgraph==0.3.8
qt-gui==0.3.8
qt-gui-cpp==0.3.8
qt-gui-py-common==0.3.8
rainbowhat==0.0.1
resource-retriever==1.12.4
robot-upstart==0.3.0
roman==2.0.0
rosapi==0.9.0
rosbag==1.12.13
rosboost-cfg==1.14.3
rosbridge-library==0.9.0
rosbridge-server==0.9.0
rosclean==1.14.3
roscreate==1.14.3
rosdep==0.12.2
rosdistro==0.6.8
rosdistro-modules==0.6.8
rosgraph==1.12.13
rosinstall==0.7.8
roslaunch==1.12.13
roslib==1.14.3
roslint==0.11.0
roslz4==1.12.13
rosmake==1.14.3
rosmaster==1.12.13
rosmsg==1.12.13
rosnode==1.12.13
rosparam==1.12.13
rospkg==1.1.4
rospkg-modules==1.1.4
rospy==1.12.13
rosservice==1.12.13
rostest==1.12.13
rostopic==1.12.13
rosunit==1.14.3
roswtf==1.12.13
RPi.GPIO==0.6.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.11
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.7
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
RTIMULib==7.2.1
rviz==1.12.16
scrollphat==0.0.7
SecretStorage==2.1.3
sense-emu==1.0
sense-hat==2.2.0
sensor-msgs==1.12.6
service-identity==16.0.0
simplejson==3.8.1
six==1.11.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.1
sn3218==1.2.6
spidev==3.0
st7036==1.4.3
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.17
tf2-kdl==0.5.17
tf2-py==0.5.17
tf2-ros==0.5.17
topic-tools==1.12.13
tornado==4.2.1
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
vcstools==0.1.40
wadllib==1.3.2
wstool==0.1.17
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.11.3
zope.interface==4.1.3
